move 'lens' tests to separate file
